requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
groq==0.9.0
transformers==4.41.0
torch==2.3.0
python-dotenv==1.0.1
twilio==9.2.2
requests==2.32.2
sqlalchemy==2.0.30
aiosqlite==0.20.0
pydantic==2.7.1
httpx==0.27.0
